Move check functions into separate modules
