+++ src/lib/OpenEXR/CMakeLists.txt
@@ -18,6 +18,7 @@ openexr_define_library(OpenEXR
     ImfCheckedArithmetic.h
     ImfChromaticities.cpp
     ImfChromaticitiesAttribute.cpp
+    ImfColorMetadata.cpp
     ImfCompositeDeepScanLine.cpp
     ImfCompression.cpp
     ImfCompression.h
@@ -136,6 +137,7 @@ openexr_define_library(OpenEXR
     ImfChannelListAttribute.h
     ImfChromaticities.h
     ImfChromaticitiesAttribute.h
+    ImfColorMetadata.h
     ImfCompositeDeepScanLine.h
     ImfCompression.h
     ImfCompressionAttribute.h
