+++ src/lib/OpenEXRCore/CMakeLists.txt
@@ -76,6 +76,7 @@ openexr_define_library(OpenEXRCore
     part.c
     part_attr.c
     std_attr.c
+    color_metadata.c
 
     parse_header.c
     write_header.c
@@ -98,6 +99,7 @@ openexr_define_library(OpenEXRCore
     openexr_base.h
     openexr_chunkio.h
     openexr_coding.h
+    openexr_color_metadata.h
     openexr_compression.h
     openexr_config.h
     openexr_context.h
