+++ src/test/OpenEXRCoreTest/CMakeLists.txt
@@ -6,8 +6,8 @@ add_executable(OpenEXRCoreTest
   base_units.h
   buffer.cpp
   buffer.h
-  color_interop.cpp
-  color_interop.h
+  color_metadata.cpp
+  color_metadata.h
   compression.cpp
   compressionTables.cpp
   compression.h
