+++ src/test/OpenEXRCoreTest/color_metadata.h
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Copyright Contributors to the OpenEXR Project.
 
-#ifndef OPENEXR_CORE_TEST_COLOR_INTEROP_H
-#define OPENEXR_CORE_TEST_COLOR_INTEROP_H
+#ifndef OPENEXR_CORE_TEST_COLOR_METADATA_H
+#define OPENEXR_CORE_TEST_COLOR_METADATA_H
 
 #include <string>
 
@@ -10,4 +10,4 @@ void testColorInteropChromaticities (const std::string& tempdir);
 void testColorMetadataValues (const std::string& tempdir);
 void testColorMetadataFile (const std::string& tempdir);
 
-#endif // OPENEXR_CORE_TEST_COLOR_INTEROP_H
+#endif // OPENEXR_CORE_TEST_COLOR_METADATA_H
