+++ src/test/OpenEXRCoreTest/main.cpp
@@ -21,7 +21,7 @@
 #endif
 
 #include "base_units.h"
-#include "color_interop.h"
+#include "color_metadata.h"
 #include "compression.h"
 #include "deep.h"
 #include "general_attr.h"
@@ -167,9 +167,9 @@ main (int argc, char* argv[])
     TEST (testAttrHandler, "gen_attr");
     TEST (testAttrLists, "gen_attr");
 
-    TEST (testColorInteropChromaticities, "color_interop");
-    TEST (testColorMetadataValues, "color_interop");
-    TEST (testColorMetadataFile, "color_interop");
+    TEST (testColorInteropChromaticities, "color_metadata");
+    TEST (testColorMetadataValues, "color_metadata");
+    TEST (testColorMetadataFile, "color_metadata");
 
     TEST (testReadBadArgs, "core_read");
     TEST (testReadBadFiles, "core_read");
